Add ignore to prepare for flow analysis change.
I will soon be landing a change to flow analysis that causes calls to
`identical` to be accounted for in type promotion, so the statement:

    return identical(result, null) ? _emptyList : result!;

will no longer need a `!`.  To avoid a nuisance
unnecessary_non_null_assertion warning when this change happens, we
add an "ignore" comment.  Once flow analysis the change has landed,
I'll make a follow-up change that removes the unnecessary `!`.